v4l: xilinx: switch: Fix warnings for OVERFLOW_BEFORE_WIDEN
This patch fix the warnings for the OVERFLOW_BEFORE_WIDEN type.

The actual warnings are:

For xvsw_s_stream():

Event  overflow_before_widen: Potentially overflowing expression "1 <<
xsw->nsources" with type "int" (32 bits, signed) is evaluated using
32-bit arithematic, and then used in a context that expects an
expression of type "unsigned long" (64 bits, unsigned).

Event remediation: To avoid overflow, cast "1" to type "unsigned long".

The checkpatch warns with this fix as below, but the coverity warning
sounds more reasonable. Hence ignoring the checkpatch warning for now.

WARNING: Unnecessary typecast of c90 int constant
+		   ((((unsigned long)1 << xsw->nsources) - 1) << 4) |

total: 0 errors, 1 warnings, 8 lines checked

State: pending

+++ drivers/media/platform/xilinx/xilinx-switch.c
@@ -104,7 +104,7 @@ static int xsw_s_stream(struct v4l2_subdev *subdev, int enable)
 	xvip_write(&xsw->xvip, XSW_CORE_CH_CTRL, routing);
 
 	xvip_write(&xsw->xvip, XVIP_CTRL_CONTROL,
-		   (((1 << xsw->nsources) - 1) << 4) |
+		   ((((unsigned long)1 << xsw->nsources) - 1) << 4) |
 		   XVIP_CTRL_CONTROL_SW_ENABLE);
 
 	return 0;
